Gcc and CMake parsing scripts are alredy defined for further development, but the main get-em script is still far from complete.
Also added the "etc/defaults.sh" script for having a ready to be accessed list of the default versions meant to be used with this version of OpenFOAM.
Nonetheless, this defaults.sh file should in fact go to OpenFOAM's main "etc/config" folder instead.

Since I added the defaults.sh, I adapted the make* scripts to use the respective default settings.